feat: call exported zig function from js

+++ app/main.zig
@@ -1,5 +1,5 @@
 pub fn main() !void {}
 
-pub export fn add(a: i64, b: i64) i64 {
+pub export fn add(a: i32, b: i32) i32 {
     return a + b;
 }
